+++ src/frontend/components/PublicPausedLiveVideo/index.spec.tsx
@@ -1,36 +1,11 @@
 import { Grommet } from 'grommet';
 import { DateTime } from 'luxon';
 import React from 'react';
-import { act, render, screen, waitFor } from '@testing-library/react';
+import { act, render, screen } from '@testing-library/react';
 
-import { getResource } from 'data/sideEffects/getResource';
-import { resumeLive } from 'utils/resumeLive';
 import { videoMockFactory } from 'utils/tests/factories';
 import { wrapInIntlProvider } from 'utils/tests/intl';
 import { PublicPausedLiveVideo } from '.';
-import { modelName } from 'types/models';
-
-jest.mock('video.js', () => ({
-  __esModule: true,
-  default: {
-    getPlayers: () => ({
-      bb8: {
-        currentSource: () => 'https://live.m3u8',
-        src: jest.fn(),
-      },
-    }),
-  },
-}));
-
-jest.mock('data/sideEffects/getResource', () => ({
-  getResource: jest.fn(),
-}));
-jest.mock('utils/resumeLive', () => ({
-  resumeLive: jest.fn(),
-}));
-
-const mockResumeLive = resumeLive as jest.MockedFunction<typeof resumeLive>;
-const mockGetResource = getResource as jest.MockedFunction<typeof getResource>;
 
 describe('PublicPausedLiveVideo', () => {
   beforeEach(() => {
@@ -53,7 +28,6 @@ describe('PublicPausedLiveVideo', () => {
       },
     });
     const videoNode = document.createElement('video');
-    mockResumeLive.mockResolvedValue();
 
     render(
       wrapInIntlProvider(
@@ -71,13 +45,11 @@ describe('PublicPausedLiveVideo', () => {
     expect(waitingSentence).toHaveTextContent(
       'The webinar is paused since 00:00:05',
     );
-    expect(mockGetResource).not.toHaveBeenCalled();
   });
 
   it('renders the component without timer', () => {
     const video = videoMockFactory();
     const videoNode = document.createElement('video');
-    mockResumeLive.mockResolvedValue();
 
     render(
       wrapInIntlProvider(
@@ -94,39 +66,6 @@ describe('PublicPausedLiveVideo', () => {
     expect(
       screen.queryByText(/the webinar is in paused since/i),
     ).not.toBeInTheDocument();
-
-    expect(mockResumeLive).toHaveBeenCalledWith(video);
-    expect(mockGetResource).not.toHaveBeenCalled();
-  });
-
-  it('calls getResource to refresh video state if resumeLive function fails', async () => {
-    const video = videoMockFactory({
-      live_info: {
-        paused_at: `${DateTime.now().minus({ seconds: 5 }).toSeconds()}`,
-      },
-    });
-    const videoNode = document.createElement('video');
-    mockResumeLive.mockRejectedValue('');
-
-    render(
-      wrapInIntlProvider(
-        <Grommet>
-          <PublicPausedLiveVideo video={video} videoNodeRef={videoNode} />
-        </Grommet>,
-      ),
-    );
-
-    screen.getByText('Webinar is paused');
-    screen.getByText(
-      'The webinar is paused. When resumed, the video will start again.',
-    );
-    const waitingSentence = screen.getByText(/the webinar is paused since/i);
-    expect(waitingSentence).toHaveTextContent(
-      'The webinar is paused since 00:00:05',
-    );
-    await waitFor(() =>
-      expect(mockGetResource).toHaveBeenCalledWith(modelName.VIDEOS, video.id),
-    );
   });
 
   it('displays day and time when live is paused since more than 24 hours', () => {
@@ -137,7 +76,6 @@ describe('PublicPausedLiveVideo', () => {
       },
     });
     const videoNode = document.createElement('video');
-    mockResumeLive.mockResolvedValue();
 
     render(
       wrapInIntlProvider(
@@ -166,7 +104,6 @@ describe('PublicPausedLiveVideo', () => {
       },
     });
     const videoNode = document.createElement('video');
-    mockResumeLive.mockResolvedValue();
 
     render(
       wrapInIntlProvider(