+++ src/frontend/types/VideoPlayer.ts
@@ -4,6 +4,8 @@ import { Video } from './tracks';
 export interface VideoPlayerInterface {
   /** Destroy the instance and garbage collect any elements. */
   destroy(): void;
+  getSource(): string;
+  setSource(url: string): void;
 }
 
 export type VideoPlayerCreator = (