+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Converse.js UI is not used anymore, react components are used instead
 - Add anonymous_id parameter to register a user to a scheduled webinar
 - Update live layout for students
+- Use websocket to update video resource instead of polling the API
 
 ### Removed